Moved the rest entites to the packtage entites
