+++ src/main/java/org/newton/webshop/models/entities/Account.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
-import org.newton.webshop.models.Customer;
 
 import javax.persistence.*;
 import java.time.LocalDate;
