+++ src/main/java/org/newton/webshop/models/entities/Cart.java
@@ -1,11 +1,10 @@
-package org.newton.webshop.models;
+package org.newton.webshop.models.entities;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
-import org.newton.webshop.models.entities.Order;
 
 import javax.persistence.*;
 import java.util.Set;
