+++ src/main/java/org/newton/webshop/models/entities/Category.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
-import org.newton.webshop.models.Product;
 
 import javax.persistence.*;
 import java.util.HashSet;
