+++ src/main/java/org/newton/webshop/models/entities/Item.java
@@ -1,10 +1,9 @@
-package org.newton.webshop.models;
+package org.newton.webshop.models.entities;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
-import org.newton.webshop.models.entities.Review;
 
 import javax.persistence.*;
