+++ src/main/java/org/newton/webshop/repositories/CartRepository.java
@@ -1,6 +1,6 @@
 package org.newton.webshop.repositories;
 
-import org.newton.webshop.models.Cart;
+import org.newton.webshop.models.entities.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Cart, String> {
