+++ src/main/java/org/newton/webshop/repositories/CustomerRepository.java
@@ -1,7 +1,6 @@
 package org.newton.webshop.repositories;
 
-import org.newton.webshop.models.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.newton.webshop.models.entities.Customer;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CustomerRepository extends CrudRepository<Customer, String> {
