+++ src/main/java/org/newton/webshop/repositories/ItemRepository.java
@@ -1,6 +1,6 @@
 package org.newton.webshop.repositories;
 
-import org.newton.webshop.models.Item;
+import org.newton.webshop.models.entities.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, String> {
