+++ src/main/java/org/newton/webshop/repositories/ProductRepository.java
@@ -1,6 +1,6 @@
 package org.newton.webshop.repositories;
 
-import org.newton.webshop.models.Product;
+import org.newton.webshop.models.entities.Product;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProductRepository extends CrudRepository<Product, String> {
